+++ client/config/contentMap.config.tsx
@@ -16,6 +16,7 @@ import TodoList from '../pages/TodoList';
 import Audio from '../pages/Audio';
 import Agent from '../pages/Agent';
 import Personalization from '../pages/Personalization';
+import Subscribe from '../pages/Subscribe';
 
 const contentMap: Record<string, JSX.Element> = {
   start: <StartPage />,
@@ -33,6 +34,7 @@ const contentMap: Record<string, JSX.Element> = {
   community: <Community />,
   todo: <TodoList />,
   agent: <Agent />,
-  personalization:<Personalization/>
+  personalization: <Personalization />,
+  subscribe: <Subscribe />,
 };
 export default contentMap;