+++ client/main.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import 'virtual:uno.css';
 import { ActiveItemProvider } from './context/activeItemContext';
 import { SettingsProvider } from './context/drawerSettingContext';
-import 'virtual:uno.css';
-
+import App from './App';
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
     <ActiveItemProvider>