+++ client/pages/Subscribe.tsx
@@ -1,0 +1,10 @@
+import PageBox from "../components/PageBox"
+
+const Subscribe:React.FC = () => {
+    return(
+        <PageBox>
+            <div></div>
+        </PageBox>
+    )
+}
+export default Subscribe;