+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 0.4.3
 1. feature: Account(include server and client)
+2. Auth
+3. Post page and Dashboard welcome words
 
 ### 0.4.2
 1. delete "Process manager" module