+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.4.4
+1. better module config page
+2. better docs page
+3. start to "task page"
 ### 0.4.3
 1. feature: Account(include server and client)
 2. Auth